+++ src/main/java/com/perfulandiaSPA/perfulandia/serviceinventario/controller/SucursalController.java
@@ -1,8 +1,14 @@
 package com.perfulandiaSPA.perfulandia.serviceinventario.controller;
 
-import com.perfulandiaSPA.perfulandia.serviceinventario.model.Producto;
 import com.perfulandiaSPA.perfulandia.serviceinventario.model.Sucursal;
 import com.perfulandiaSPA.perfulandia.serviceinventario.service.SucursalService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,12 +17,20 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/sucursales")
+@Tag(name="Sucursales", description = "Operaciones relacionadas con los sucursales")
 public class SucursalController {
 
     @Autowired
     private SucursalService sucursalService;
 
     @GetMapping()
+    @Operation(summary = "Obtener todas las sucursales", description = "Obtiene una lista de todas las sucursales")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Lista de sucursales obtenida exitosamente",
+                    content = @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Sucursal.class))),
+            @ApiResponse(responseCode = "204", description = "No hay sucursales disponibles")
+    })
     public ResponseEntity<List<Sucursal>> listar() {
         List<Sucursal> sucursales = sucursalService.findAll();
         if (sucursales.isEmpty()) {
@@ -26,13 +40,30 @@ public class SucursalController {
     }
 
     @GetMapping("/{id}")
+    @Operation(summary = "Obtener sucursal por ID", description = "Retorna la información de una sucursal específico dado su ID")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Sucursal obtenida exitosamente",
+                    content = @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Sucursal.class))),
+            @ApiResponse(responseCode = "404", description = "Sucursal no encontrada")
+    })
     public ResponseEntity<Sucursal> obtenerPorID(@PathVariable Long id) {
         Sucursal sucursal = sucursalService.findById(id);
+        if (sucursal == null) {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok(sucursal);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Sucursal> actualizarSucursal(@PathVariable Long id, @RequestBody Sucursal sucursal) {
+    @Operation(summary = "Actualizar una sucursal", description = "Actualiza una sucursal existente")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Sucursal actualizada exitosamente",
+                    content = @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Sucursal.class))),
+            @ApiResponse(responseCode = "404", description = "Sucursal no encontrada")
+    })
+    public ResponseEntity<Sucursal> actualizarSucursal(@PathVariable Long id, @Valid @RequestBody Sucursal sucursal) {
         Sucursal existente = sucursalService.findById(id);
         if (existente == null) {
             return ResponseEntity.notFound().build();
@@ -42,13 +73,29 @@ public class SucursalController {
     }
 
     @PostMapping
-    public ResponseEntity<Sucursal> crearSucursal(@RequestBody Sucursal sucursal) {
+    @Operation(summary = "Crear una nueva sucursal", description = "Crea una sucursal y retorna la sucursal creada con su ID generado")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Sucursal creada exitosamente",
+                    content = @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Sucursal.class))),
+            @ApiResponse(responseCode = "400", description = "Datos proporcionados inválidos")
+    })
+    public ResponseEntity<Sucursal> crearSucursal(@Valid @RequestBody Sucursal sucursal) {
         Sucursal nuevo = sucursalService.save(sucursal);
         return new ResponseEntity<>(nuevo, HttpStatus.CREATED);
     }
 
     @DeleteMapping("/{id}")
+    @Operation(summary = "Eliminar una sucursal", description = "Elimina una sucursal existente dado su ID")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "Sucursal eliminada exitosamente"),
+            @ApiResponse(responseCode = "404", description = "Sucursal no encontrada")
+    })
     public ResponseEntity<Void> eliminarSucursal(@PathVariable Long id) {
+        Sucursal sucursal = sucursalService.findById(id);
+        if (sucursal == null) {
+            return ResponseEntity.notFound().build();
+        }
         sucursalService.delete(id);
         return ResponseEntity.noContent().build();
     }
