+++ src/main/java/com/perfulandiaSPA/perfulandia/servicepedido/controller/PedidoController.java
@@ -2,6 +2,13 @@ package com.perfulandiaSPA.perfulandia.servicepedido.controller;
 
 import com.perfulandiaSPA.perfulandia.servicepedido.model.Pedido;
 import com.perfulandiaSPA.perfulandia.servicepedido.service.PedidoService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,12 +17,20 @@ import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/pedidos")
+@Tag(name="Pedidos", description = "Operaciones relacionadas con los pedidos")
 public class PedidoController {
 
     @Autowired
     private PedidoService pedidoService;
 
     @GetMapping()
+    @Operation(summary = "Obtener todos los pedidos", description = "Obtiene una lista de todos los pedidos")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Lista de pedidos obtenida exitosamente",
+                    content = @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Pedido.class))),
+            @ApiResponse(responseCode = "204", description = "No hay pedidos disponibles")
+    })
     public ResponseEntity<List<Pedido>> listar() {
         List<Pedido> pedidos = pedidoService.findAll();
         if (pedidos.isEmpty()) {
@@ -25,13 +40,30 @@ public class PedidoController {
     }
 
     @GetMapping("/{id}")
+    @Operation(summary = "Obtener pedido por ID", description = "Retorna la información de un pedido específico dado su ID")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Pedido obtenido exitosamente",
+                    content = @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Pedido.class))),
+            @ApiResponse(responseCode = "404", description = "Pedido no encontrado")
+    })
     public ResponseEntity<Pedido> obtenerPorID(@PathVariable Long id) {
         Pedido pedido = pedidoService.findById(id);
+        if (pedido == null) {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok(pedido);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Pedido> actualizarPedido(@PathVariable Long id, @RequestBody Pedido pedido) {
+    @Operation(summary = "Actualizar un pedido", description = "Actualiza un pedido existente")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Pedido actualizado exitosamente",
+                    content = @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Pedido.class))),
+            @ApiResponse(responseCode = "404", description = "Pedido no encontrada")
+    })
+    public ResponseEntity<Pedido> actualizarPedido(@PathVariable Long id, @Valid @RequestBody Pedido pedido) {
         Pedido existente = pedidoService.findById(id);
         if (existente == null) {
             return ResponseEntity.notFound().build();
@@ -41,13 +73,29 @@ public class PedidoController {
     }
 
     @PostMapping
-    public ResponseEntity<Pedido> crearPedido(@RequestBody Pedido pedido) {
+    @Operation(summary = "Crear un nuevo pedido", description = "Crea un pedido y retorna el pedido creado con su ID generado")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "Pedido creado exitosamente",
+                    content = @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = Pedido.class))),
+            @ApiResponse(responseCode = "400", description = "Datos proporcionados inválidos")
+    })
+    public ResponseEntity<Pedido> crearPedido(@Valid @RequestBody Pedido pedido) {
         Pedido nuevo = pedidoService.save(pedido);
         return new ResponseEntity<>(nuevo, HttpStatus.CREATED);
     }
 
     @DeleteMapping("/{id}")
+    @Operation(summary = "Eliminar un pedido", description = "Elimina un pedido existente dado su ID")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "Pedido eliminado exitosamente"),
+            @ApiResponse(responseCode = "404", description = "Pedido no encontrado")
+    })
     public ResponseEntity<Void> eliminarPedido(@PathVariable Long id) {
+        Pedido pedido = pedidoService.findById(id);
+        if (pedido == null) {
+            return ResponseEntity.notFound().build();
+        }
         pedidoService.delete(id);
         return ResponseEntity.noContent().build();
     }
